implement review suggestions
