+++ packages/blitz/src/global.ts
@@ -1,14 +1,8 @@
-import {
-  BeforeHttpRequest,
-  BeforeHttpResponse,
-  OnRpcError,
-  EventHook,
-  OnSessionCreated,
-} from "./types"
+import {BeforeHttpRequest, BeforeHttpResponse} from "./types"
 declare global {
   var _blitz_prismaClient: any
-  var __BLITZ_beforeHttpRequest: BeforeHttpRequest
-  var __BLITZ_beforeHttpResponse: BeforeHttpResponse
-  var __BLITZ_onRpcError: EventHook<OnRpcError>
-  var __BLITZ_onSessionCreated: EventHook<OnSessionCreated>
+  var __BLITZ_MIDDLEWARE_HOOKS: {
+    beforeHttpRequest: BeforeHttpRequest
+    beforeHttpResponse: BeforeHttpResponse
+  }
 }
