+++ packages/blitz/src/utils/index.ts
@@ -24,6 +24,15 @@ function toInteger(dirtyNumber: unknown) {
   return number < 0 ? Math.ceil(number) : Math.floor(number)
 }
 
+// plugin utils
+export function merge<T, U>([...fns]: Array<(args: T) => U>) {
+  return (args: T) => fns.map((fn) => fn(args))
+}
+
+export function pipe<T>(...fns: Array<(x: T) => T>) {
+  return (x: T) => fns.reduce((v, f) => f(v), x)
+}
+
 export const isPast = (date: Date) => {
   return date.getTime() < Date.now()
 }
